requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow
scipy>=1.4.1
torch==1.8.0
torchvision==0.9.0
tqdm>=4.41.0